Add bulk "event type" to run the tagging on all movies in radarr
This action should be done on an environment that has access to the
movie files at the same location as the radarr process.
Either run it where radarr runs, or use the same jail mountpoints or
container path mappings or whatever.

`./tag_dvfelmel.sh bulk`

This command will ask radarr for all movies, filter out the ones
without a file, and run the script on them as if radarr connect
triggered the script